Keep 2.x extensions codestarts embeded in tooling and mark them legacy
