--- independent-projects/tools/base-codestarts/src/main/resources/codestarts/quarkus/extension-codestarts/resteasy-codestart/kotlin/src/main/kotlin/org/acme/{resource.class-name}.tpl.qute.kt
@@ -1,14 +1,0 @@
-package org.acme
-
-import jakarta.ws.rs.GET
-import jakarta.ws.rs.Path
-import jakarta.ws.rs.Produces
-import jakarta.ws.rs.core.MediaType
-
-@Path("{resource.path}")
-class {resource.class-name} {
-
-    @GET
-    @Produces(MediaType.TEXT_PLAIN)
-    fun hello() = "{resource.response}"
-}
