--- independent-projects/tools/base-codestarts/src/main/resources/codestarts/quarkus/extension-codestarts/resteasy-reactive-codestart/scala/src/main/scala/org/acme/{resource.class-name}.tpl.qute.scala
@@ -1,12 +1,0 @@
-package org.acme
-
-import jakarta.ws.rs.\{GET, Path, Produces}
-import jakarta.ws.rs.core.MediaType
-
-@Path("{resource.path}")
-class {resource.class-name} {
-
-    @GET
-    @Produces(Array[String](MediaType.TEXT_PLAIN))
-    def hello() = "{resource.response}"
-}
