+++ independent-projects/tools/base-codestarts/src/main/resources/codestarts/quarkus/legacy-extensions-codestarts/README.md
@@ -1,0 +1,2 @@
+Those extensions codestarts are used with Quarkus platform 2.10 and below.
+From 2.11 and onward, they are provided by the platform.
