+++ independent-projects/tools/base-codestarts/src/main/resources/codestarts/quarkus/legacy-extensions-codestarts/resteasy-codestart/kotlin/src/main/kotlin/org/acme/{resource.class-name}.tpl.qute.kt
@@ -1,14 +1,14 @@
-package ilove.quark.us
+package org.acme
 
 import javax.ws.rs.GET
 import javax.ws.rs.Path
 import javax.ws.rs.Produces
 import javax.ws.rs.core.MediaType
 
-@Path("/hello")
-class GreetingResource {
+@Path("{resource.path}")
+class {resource.class-name} {
 
     @GET
     @Produces(MediaType.TEXT_PLAIN)
-    fun hello() = "Hello from RESTEasy Reactive"
+    fun hello() = "{resource.response}"
 }
