+++ independent-projects/tools/base-codestarts/src/main/resources/codestarts/quarkus/legacy-extensions-codestarts/resteasy-codestart/scala/src/main/scala/org/acme/{resource.class-name}.tpl.qute.scala
@@ -1,0 +1,12 @@
+package org.acme
+
+import javax.ws.rs.\{GET, Path, Produces}
+import javax.ws.rs.core.MediaType
+
+@Path("{resource.path}")
+class {resource.class-name} {
+
+    @GET
+    @Produces(Array[String](MediaType.TEXT_PLAIN))
+    def hello() = "{resource.response}"
+}
