+++ independent-projects/tools/base-codestarts/src/main/resources/codestarts/quarkus/legacy-extensions-codestarts/spring-web-codestart/base/README.tpl.qute.md
