+++ independent-projects/tools/base-codestarts/src/main/resources/codestarts/quarkus/legacy-extensions-codestarts/spring-web-codestart/base/src/main/resources/META-INF/resources/index.entry.qute.html
