+++ independent-projects/tools/base-codestarts/src/main/resources/codestarts/quarkus/legacy-extensions-codestarts/spring-web-codestart/java/src/native-test/java/org/acme/{resource.class-name}IT.tpl.qute.java
