+++ independent-projects/tools/base-codestarts/src/main/resources/codestarts/quarkus/legacy-extensions-codestarts/spring-web-codestart/java/src/test/java/org/acme/{resource.class-name}Test.tpl.qute.java
