+++ independent-projects/tools/base-codestarts/src/main/resources/codestarts/quarkus/legacy-extensions-codestarts/spring-web-codestart/kotlin/src/main/kotlin/org/acme/{resource.class-name}.tpl.qute.kt
