+++ independent-projects/tools/base-codestarts/src/main/resources/codestarts/quarkus/legacy-extensions-codestarts/spring-web-codestart/scala/src/native-test/scala/org/acme/{resource.class-name}IT.tpl.qute.scala
