+++ independent-projects/tools/base-codestarts/src/main/resources/codestarts/quarkus/legacy-extensions-codestarts/spring-web-codestart/scala/src/test/scala/org/acme/{resource.class-name}Test.tpl.qute.scala
